+++ Makefile
@@ -4,9 +4,6 @@ install deps:
 brain-games:
 	node bin/brain-games.js
 
-brain-even:
-	node bin/brain-even.js
-
 publish:
 	npm publish --dry-run
